permission validation for delete

+++ backend/src/resolvers/Mutation.js
@@ -45,7 +45,16 @@ const mutations = {
   },
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
-    const item = await ctx.db.query.item({ where }, `{ id title }`);
+    const item = await ctx.db.query.item({ where }, `{ id title user { id } }`);
+
+    const owmsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission => {
+      const validPermissions = ['ADMIN', 'ITEMDELETE'];
+      return validPermissions.includes(permission);
+    });
+
+    if (!owmsItem && !hasPermission) throw new Error(`You don't have permission to do that!`);
+
     return ctx.db.mutation.deleteItem({ where }, info);
   },
   async signup(parent, args, ctx, info) {
